google/requirements.txt
requests==2.19.1
google-api-core==1.2.1
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-cloud-vision==0.32.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio
oauth2client==4.1.2
protobuf==3.6.0
Pillow
python-Levenshtein
fuzzywuzzy